Initial commit of jekyll changes
Lots of stuff in this commit. The base idea was to use Jekyll to help with design and flexibilitly. I have a home page with greeting logo and the deck embedded. I have a nav bar with home, deck, and faq pages. The deck page has the pdf displayed, and the faq has the FAQ md text.